Split common/ Types.ts into .d.ts and Constants.ts
