+++ src/common/buffer/Types.d.ts
