+++ src/renderer/BaseRenderLayer.ts
@@ -6,7 +6,8 @@
 import { IRenderLayer } from './Types';
 import { IRenderDimensions } from 'browser/renderer/Types';
 import { ITerminal } from '../Types';
-import { ICellData, DEFAULT_COLOR } from 'common/Types';
+import { ICellData } from 'common/Types';
+import { DEFAULT_COLOR } from 'common/buffer/Constants';
 import { DIM_OPACITY, INVERTED_DEFAULT_COLOR, IGlyphIdentifier } from './atlas/Types';
 import { BaseCharAtlas } from './atlas/BaseCharAtlas';
 import { acquireCharAtlas } from './atlas/CharAtlasCache';
